correct event name and move into logging handler
