+++ src/main/java/io/cryostat/net/web/WebServer.java
@@ -68,7 +68,6 @@ import io.cryostat.net.web.http.api.v2.ApiException;
 import io.cryostat.util.HttpStatusCodeIdentifier;
 
 import com.google.gson.Gson;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.Handler;
 import io.vertx.core.http.HttpHeaders;
@@ -77,10 +76,6 @@ import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.HttpException;
 import io.vertx.ext.web.impl.BlockingHandlerDecorator;
-import jdk.jfr.Category;
-import jdk.jfr.Event;
-import jdk.jfr.Label;
-import jdk.jfr.Name;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.http.client.utils.URIBuilder;
 import org.apache.http.impl.EnglishReasonPhraseCatalog;
@@ -239,52 +234,7 @@ public class WebServer extends AbstractVerticle {
                     }
                 });
 
-        this.server.requestHandler(
-                req -> {
-                    WebServerRequest evt =
-                            new WebServerRequest(
-                                    req.remoteAddress().host(),
-                                    req.remoteAddress().port(),
-                                    req.method().toString(),
-                                    req.path());
-                    evt.begin();
-
-                    req.response()
-                            .endHandler(
-                                    (res) -> {
-                                        evt.setStatusCode(req.response().getStatusCode());
-                                        evt.end();
-                                        if (evt.shouldCommit()) {
-                                            evt.commit();
-                                        }
-                                    });
-                    router.handle(req);
-                });
-    }
-
-    @Name("io.cryostat.net.web.WebServer.WebServerRequest")
-    @Label("Web Server Request")
-    @Category("Cryostat")
-    @SuppressFBWarnings(
-            value = "URF_UNREAD_FIELD",
-            justification = "The event fields are recorded with JFR instead of accessed directly")
-    public static class WebServerRequest extends Event {
-        String host;
-        int port;
-        String method;
-        String path;
-        int statusCode;
-
-        public WebServerRequest(String host, int port, String method, String path) {
-            this.host = host;
-            this.port = port;
-            this.method = method;
-            this.path = path;
-        }
-
-        public void setStatusCode(int code) {
-            this.statusCode = code;
-        }
+        this.server.requestHandler(router::handle);
     }
 
     @Override
